requirements.txt
appnope==0.1.3
APScheduler==3.6.3
asttokens==2.0.5
backcall==0.2.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
debugpy==1.6.2
decorator==5.1.1
dialogflow==2.0.0
entrypoints==0.4
executing==0.9.1
feedparser==6.0.10
Flask==2.1.3
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.32.0
google-auth==1.35.0
google-cloud-dialogflow==2.15.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
requests==2.28.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
stack-data==0.3.0
telegram==0.0.1
tornado==6.2
traitlets==5.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.0
zipp==3.8.1
